hibernate zintegrowany z poprzednimi zadaniami

+++ src/main/java/com/umcsuser/carrent/repositories/VehicleRepository.java
@@ -2,12 +2,13 @@ package com.umcsuser.carrent.repositories;
 
 import com.umcsuser.carrent.models.Vehicle;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
 public interface VehicleRepository {
-    void addVehicle(String id, String category, String brand, String model, int year, String plate, int price, Map<String, Object> atributes);
+    void addVehicle(String id, String category, String brand, String model, int year, String plate, BigDecimal price, Map<String, Object> atributes);
 
     void removeVehicle(String id);
